Fix cached user issue (#4)
If a user logs out, then logs in as a new user, we weren't checking to see if the user had changed, only that we had cookies so we would use the cached used until the cache expired in 15 minutes. Now we verify the cached username matches the cookie username and refetch if they don't.